v1.8: show each external disk as E; DA hot-plug; Swift 6.3.3 compile
Fixed+External PCI-E/USB volumes were classified internal because
kind required ejectable. Use DA DeviceInternal + USB/Thunderbolt
protocol. Bar is D (hottest system) plus one E per external so Data
cannot hide expansion disks. DiskArbitration appear/disappear/path
callbacks refresh immediately. Compile: contentRect/frameRect are
funcs, @MainActor on QuotaBarApp, PROC_ALL_PIDS as UInt32, MAXPATHLEN.

+++ QuotaBar/UsageSource.swift
@@ -122,11 +122,11 @@ enum ProcessProbe {
     }
 
     private static func scanClaude() -> Bool {
-        let raw = proc_listpids(PROC_ALL_PIDS, 0, nil, 0)
+        let raw = proc_listpids(UInt32(PROC_ALL_PIDS), 0, nil, 0)
         guard raw > 0 else { return false }
         let count = Int(raw) / MemoryLayout<pid_t>.stride
         var pids = [pid_t](repeating: 0, count: count)
-        let filled = proc_listpids(PROC_ALL_PIDS, 0, &pids, raw)
+        let filled = proc_listpids(UInt32(PROC_ALL_PIDS), 0, &pids, raw)
         guard filled > 0 else { return false }
         let n = Int(filled) / MemoryLayout<pid_t>.stride
         for pid in pids.prefix(n) where pid > 0 {
